Add randomString() to AuthUtils
