+++ lib/authorizer/digest.js
@@ -3,6 +3,7 @@ var _ = require('lodash'),
     urlEncoder = require('postman-url-encoder'),
     RequestBody = require('postman-collection').RequestBody,
     bodyBuilder = require('../requester/core-body-builder'),
+    AuthUtils = require('./util'),
 
     EMPTY = '',
     ONE = '00000001',
@@ -17,8 +18,6 @@ var _ = require('lodash'),
     AUTH_INT = 'auth-int',
     AUTHORIZATION = 'Authorization',
     MD5_SESS = 'MD5-sess',
-    ASCII_SOURCE = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789',
-    ASCII_SOURCE_LENGTH = ASCII_SOURCE.length,
     USERNAME_EQUALS_QUOTE = 'username="',
     REALM_EQUALS_QUOTE = 'realm="',
     NONCE_EQUALS_QUOTE = 'nonce="',
@@ -97,26 +96,6 @@ if (!_.includes(crypto.getHashes(), 'sha512-256')) {
         }
     });
 }
-
-/**
- * Generates a random string of given length
- *
- * @todo Move this to util.js. After moving use that for hawk auth too
- * @param {Number} length
- */
-function randomString (length) {
-    length = length || 6;
-
-    var result = [],
-        i;
-
-    for (i = 0; i < length; i++) {
-        result[i] = ASCII_SOURCE[(Math.random() * ASCII_SOURCE_LENGTH) | 0];
-    }
-
-    return result.join(EMPTY);
-}
-
 
 /**
  * Extracts a Digest Auth field from a WWW-Authenticate header value using a given regexp.
@@ -321,7 +300,7 @@ module.exports = {
             qop && (authParams.qop = qop);
 
             if (authParams.qop || auth.get(QOP)) {
-                authParams.clientNonce = randomString(8);
+                authParams.clientNonce = AuthUtils.randomString(8);
                 authParams.nonceCount = ONE;
             }
