+++ lib/authorizer/hawk.js
@@ -5,29 +5,9 @@ var url = require('url'),
     RequestBody = require('postman-collection').RequestBody,
     bodyBuilder = require('../requester/core-body-builder'),
     urlEncoder = require('postman-url-encoder'),
+    AuthUtils = require('./util'),
 
-    ASCII_SOURCE = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789',
-    ASCII_SOURCE_LENGTH = ASCII_SOURCE.length,
-    AUTHORIZATION = 'Authorization',
-    EMPTY = '';
-
-/**
- * Generates a random string of given length (useful for nonce generation, etc).
- *
- * @param {Number} length
- */
-function randomString (length) {
-    length = length || 6;
-
-    var result = [],
-        i;
-
-    for (i = 0; i < length; i++) {
-        result[i] = ASCII_SOURCE[(Math.random() * ASCII_SOURCE_LENGTH) | 0];
-    }
-
-    return result.join(EMPTY);
-}
+    AUTHORIZATION = 'Authorization';
 
 /**
  * Calculates body hash with given algorithm and digestEncoding.
@@ -151,7 +131,7 @@ module.exports = {
      * @param {AuthHandlerInterface~authPreHookCallback} done
      */
     pre: function (auth, done) {
-        !auth.get('nonce') && auth.set('nonce', randomString(6));
+        !auth.get('nonce') && auth.set('nonce', AuthUtils.randomString(6));
         !_.parseInt(auth.get('timestamp')) && auth.set('timestamp', Math.floor(Date.now() / 1e3));
         done(null, true);
     },
